+++ pkg/objstore/metrics_test.go
